* 2 opcode fixes by dave, thx * correced movement packets, server will no longer drop connection when moving * attempted to fix crash in openGL mode in m2 model loader * fixed SCPDatabase::GetFieldByString(), thx dave * added scp files for 3.0.9 client

+++ src/Client/World/MovementMgr.cpp
@@ -34,7 +34,7 @@ void MovementMgr::_BuildPacket(uint16 opcode)
 {
     WorldPacket *wp = new WorldPacket(opcode,4+1+4+12); // it can be larger, if we are jumping, on transport or swimming
     *wp << _moveFlags;
-    *wp << (uint8)0; // unk
+    *wp << (uint16)0; // unk
     *wp << getMSTime();
     *wp << _mychar->GetPosition();
     // TODO: transport not yet handled/done
